Mis avances seguros en el repo nuevo

+++ Application.Core/Domain/Entities/SysRegistroModificacionesJson.cs
@@ -5,9 +5,12 @@ namespace Application.Core.Domain.Entities;
 
 public partial class SysRegistroModificacionesJson
 {
-    public int IdLog { get; set; }
+    // Es la llave primaria y foránea al mismo tiempo (Relación 1 a 1)
+    public long IdLog { get; set; }
 
+    // Aquí guardaremos el JSON, pero para la base de datos es solo un texto largo
     public string? DatosAnteriores { get; set; }
 
+    // La conexión con el papá (SysRegistroModificacione)
     public virtual SysRegistroModificacione IdLogNavigation { get; set; } = null!;
 }
